insights dashboard refactor
