--- frontend/src/component/executiveDashboard/TimeToProductionChart/TimeToProductionChart.tsx
@@ -1,28 +1,0 @@
-import { type VFC } from 'react';
-import 'chartjs-adapter-date-fns';
-import { ExecutiveSummarySchema } from 'openapi';
-import { LineChart } from '../LineChart/LineChart';
-import { useProjectChartData } from '../useProjectChartData';
-
-interface IFlagsProjectChartProps {
-    projectFlagTrends: ExecutiveSummarySchema['projectFlagTrends'];
-}
-
-export const TimeToProductionChart: VFC<IFlagsProjectChartProps> = ({
-    projectFlagTrends,
-}) => {
-    const data = useProjectChartData(projectFlagTrends);
-
-    return (
-        <LineChart
-            data={data}
-            isLocalTooltip
-            overrideOptions={{
-                parsing: {
-                    yAxisKey: 'timeToProduction',
-                    xAxisKey: 'date',
-                },
-            }}
-        />
-    );
-};
