+++ frontend/src/component/executiveDashboard/components/Gauge/Gauge.tsx
