+++ frontend/src/component/executiveDashboard/componentsChart/MetricsSummaryChart/MetricsSummaryChart.tsx
@@ -1,0 +1,17 @@
+import { type VFC } from 'react';
+import 'chartjs-adapter-date-fns';
+import { ExecutiveSummarySchema } from 'openapi';
+import { LineChart } from '../../components/LineChart/LineChart';
+import { useMetricsSummary } from '../../hooks/useMetricsSummary';
+
+interface IMetricsSummaryChartProps {
+    metricsSummaryTrends: ExecutiveSummarySchema['metricsSummaryTrends'];
+}
+
+export const MetricsSummaryChart: VFC<IMetricsSummaryChartProps> = ({
+    metricsSummaryTrends,
+}) => {
+    const data = useMetricsSummary(metricsSummaryTrends, 'total');
+
+    return <LineChart data={data} />;
+};
