+++ frontend/src/component/executiveDashboard/componentsStat/UserStats/UserDistributionInfo.tsx
