+++ frontend/src/component/executiveDashboard/hooks/useFilteredFlagsSummary.ts
@@ -1,6 +1,7 @@
 import { useMemo } from 'react';
 import { ExecutiveSummarySchemaProjectFlagTrendsItem } from 'openapi';
 
+// NOTE: should we move project filtering to the backend?
 export const useFilteredFlagsSummary = (
     filteredProjectFlagTrends: ExecutiveSummarySchemaProjectFlagTrendsItem[],
 ) =>
@@ -41,6 +42,8 @@ export const useFilteredFlagsSummary = (
         return {
             ...sum,
             averageUsers,
-            averageHealth: ((sum.active / (sum.total || 1)) * 100).toFixed(0),
+            averageHealth: sum.total
+                ? ((sum.active / (sum.total || 1)) * 100).toFixed(0)
+                : undefined,
         };
     }, [filteredProjectFlagTrends]);
